+++ frontend/src/assets/fonts/Marianne/Marianne-Bold_Italic.woff2
