+++ frontend/src/assets/fonts/Marianne/Marianne-Medium_Italic.woff2
