+++ frontend/src/assets/fonts/Marianne/Marianne-Regular_Italic.woff2
